+++ Mage.Sets/src/mage/cards/m/MausoleumHarpy.java
@@ -5,7 +5,7 @@ import mage.abilities.common.DiesCreatureTriggeredAbility;
 import mage.abilities.condition.common.CitysBlessingCondition;
 import mage.abilities.decorator.ConditionalInterveningIfTriggeredAbility;
 import mage.abilities.effects.common.counter.AddCountersSourceEffect;
-import mage.abilities.hint.CitysBlessingHint;
+import mage.abilities.hint.common.CitysBlessingHint;
 import mage.abilities.keyword.AscendAbility;
 import mage.abilities.keyword.FlyingAbility;
 import mage.cards.CardImpl;
