+++ Mage/src/main/java/mage/abilities/hint/common/DeliriumHint.java
@@ -1,8 +1,10 @@
-package mage.abilities.hint;
+package mage.abilities.hint.common;
 
 import mage.abilities.Ability;
 import mage.abilities.condition.common.DeliriumCondition;
 import mage.abilities.dynamicvalue.common.CardTypesInGraveyardCount;
+import mage.abilities.hint.ConditionHint;
+import mage.abilities.hint.Hint;
 import mage.game.Game;
 
 /**
